working google oauth
Currently working with my client api key, so you'll have to email
me for the key if you would like to test. The key cannot be tracked
by git or anything like that. Right now it doesn't seem like ewu
emails work with google, but as soon as they do, we are pretty much
set. Hopefully @CountryNerd can help us with the google connection.
#12