working google oauth reciever
Now 'catches' the oauth and decodes code sent by google,
recieving the JWT with all the data we need: email, name,
etc. TODO: get a JWT parser to get at all the data, and
search the database for the user. We will also need to
ensure that EWU emails will work for google oauth.
Could we get some verification on this? #12